requirements.txt
Flask==1.1.2
mysql-connector-python==8.0.18
ulid-py==0.0.9